requirements.txt
streamlit
langchain
openai
requests
python-dotenv
tiktoken
langchain-community